requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile '.\requirements.in'
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via click
fastapi==0.115.10
    # via -r .\requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
psycopg2==2.9.10
    # via -r .\requirements.in
pydantic==2.10.6
    # via
    #   -r .\requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
python-dotenv==1.0.1
    # via -r .\requirements.in
requests==2.32.3
    # via -r .\requirements.in
sniffio==1.3.1
    # via anyio
starlette==0.46.0
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via -r .\requirements.in
